Well, there is some fundamental issues with this feature: the recover
of a FES from the acceleration of a bigger set of variables (a higher
dimensional TAMD FES) is not straightforward.

This feature branch merge back to develop discarded since the code
evolve a lot in others levels beyond the pair potential feature itself.

BTW: This branch will not be discarded yet. We have some ideas to work
in the future to solve the fundamental issues mentioned above.

+++ src/catbinsp.c
@@ -11,7 +11,7 @@ int main ( int argc, char * argv[] ) {
   int n,i,ln,timestep;
   int reading;
   double * data[4];
-  int * intdata;
+  long * intdata;
   char flag[4];
 
   for (i=1;i<argc;i++) {
@@ -34,7 +34,7 @@ int main ( int argc, char * argv[] ) {
   }
 
   for (i=0;i<2;i++) data[i]=(double*)calloc(n,sizeof(double));
-  intdata=(int*)calloc(n,sizeof(int));
+  intdata=(long*)calloc(n,sizeof(long));
 
   reading=1;
   ln=0;
@@ -42,7 +42,7 @@ int main ( int argc, char * argv[] ) {
     reading=0; 
     fread(&timestep,sizeof(int),1,fp);
     if (fileOutputLevel & 1) reading=fread(data[0],sizeof(double),n,fp);
-    if (fileOutputLevel & 2) reading=fread(intdata,sizeof(int),n,fp);
+    if (fileOutputLevel & 2) reading=fread(intdata,sizeof(long),n,fp);
 
     if (desiredOutputLevel & 1) {
       fprintf(stdout,"KNOTS    %i ",timestep);
@@ -51,7 +51,7 @@ int main ( int argc, char * argv[] ) {
     } 
     if (desiredOutputLevel & 2) {
       fprintf(stdout,"HITS     %i ",timestep);
-      for (i=0;i<n;i++) fprintf(stdout,"% 15i",intdata[i]);
+      for (i=0;i<n;i++) fprintf(stdout,"% 15li",intdata[i]);
       fprintf(stdout,"\n");
     } 
     ln++;
